dayspa_backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.1.3
django-cors-headers==4.6.0
django-unfold==0.41.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
